+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 ### Version 13.8
 - Fix incorrect scaling in the formula making
 - Fix incorrect total unit shown if above 1000g
+- Added improved ml2l functions for the PV Making app api
+- Removed links fro Android PV Making app
+- Improved PV Making app API
 
 ### Version 13.7
 - Prevent add formula for making if ingredients missing from the database